aider: Add delay before moving cursor to end of text in `tabHandler` function.
# Aider chat conversation:

USER: there is a problem with this code. when it adds blocks in the `tabHandler` it is supposed to then move the cursor to the end of the text. while the command is right it is run too soon, it takes wordpress a short time before it is ready. i need to add a delay before the commands to move the cursor to the end. a half second should be enough.

ASSISTANT: To address the issue of the cursor movement command running too soon, we need to modify the `autosuggestai.js` file to add a delay before the cursor movement commands in the `tabHandler` function. We will use `setTimeout` to introduce the half-second delay.

Here is the *SEARCH/REPLACE* block:

autosuggestai.js
```javascript
<<<<<<< SEARCH
        wp.data.dispatch('core/block-editor').selectionChange(lastBlockId, "content", parts[parts.length - 1].length, parts[parts.length - 1].length);
=======
        setTimeout(() => {
            wp.data.dispatch('core/block-editor').selectionChange(lastBlockId, "content", parts[parts.length - 1].length, parts[parts.length - 1].length);
        }, 500);
>>>>>>> REPLACE
```

This change wraps the existing `selectionChange` dispatch call within a `setTimeout` function with a delay of 500 milliseconds (half a second).